backend-django/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djoser==1.3.3
django-model-utils==3.1.2
hashids==1.2.0
requests==2.20.1
django-rest-swagger==2.1.2
django-filter==2.0.0
xmltodict==0.11.0
uWSGI==2.0.17.1
psycopg2==2.8.5
fabric==2.4.0
cryptography==2.4.2 # Disable weird Paramiko Deprecation warnings.
fcm-django==0.2.21
celery==5.2.7
redis==4.3.4
pyopendota==0.2.8